osd: fix osd_can_fit return type

+++ src/driver/osd/displayport.h
@@ -11,6 +11,6 @@ void displayport_intro();
 bool displayport_clear_async();
 osd_system_t displayport_check_system();
 
-uint8_t displayport_can_fit();
+uint32_t displayport_can_fit();
 bool displayport_push_string(uint8_t attr, uint8_t x, uint8_t y, const uint8_t *data, uint8_t size);
 bool displayport_flush();
