+++ src/driver/osd/max7456.h
@@ -49,7 +49,7 @@ void max7456_intro();
 bool max7456_clear_async();
 osd_system_t max7456_check_system();
 
-uint8_t max7456_can_fit();
+uint32_t max7456_can_fit();
 bool max7456_push_string(uint8_t attr, uint8_t x, uint8_t y, const uint8_t *data, uint8_t size);
 bool max7456_flush();
